Change value of playlist to int

+++ src/main/java/CS1530/RottenApples/models/Playlist.java
@@ -8,7 +8,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @Document(collection = "Playlists")
 public class Playlist {
     @Id
-    private Long id;
+    private int id;
     private Movie movie;
     private User user;
     private String playlistTitle;
@@ -23,7 +23,7 @@ public class Playlist {
         
     }
 
-    public Playlist(String playlistTitle, Long id) {
+    public Playlist(String playlistTitle, int id) {
         this.playlistTitle = playlistTitle;
         this.id = id;
     }
@@ -54,11 +54,11 @@ public class Playlist {
        this.isPrivate = isPrivate;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
